--- nvim/.config/nvim/lua/jsmith/init.lua
@@ -1,5 +1,0 @@
-require("jsmith.packer")
-require("jsmith.opts")
-require("jsmith.remaps")
-require("jsmith.neogit")
-
