+++ nvim/.config/nvim/lua/plugins/git.lua
